finish examextra

+++ exam1extra/src/examextra_problem1.h
@@ -4,8 +4,8 @@
 int average(int a, int b);
 float average(int a, float b);
 float average(float a, int b);
-float average(double a, int b);
-float average(int a, double b);
+double average(double a, int b);
+double average(int a, double b);
 double average(float a, double b);
 double average(double a, double b);
 #endif // EXAMEXTRA_PROBLEM1_H
